requirements.txt
# Core dependencies
pandas>=2.0.0
pyarrow>=12.0.0
openpyxl>=3.1.0
pyyaml>=6.0
jinja2>=3.1.0
click>=8.1.0
jsonschema>=4.17.0

# Optional for very large datasets
dask[dataframe]>=2023.5.0

# Progress bars
tqdm>=4.65.0

# Date parsing
python-dateutil>=2.8.0

# For colored terminal output
colorama>=0.4.6
